+++ src/test/java/GetOrderClientByNumberTest.java
@@ -8,27 +8,27 @@ import org.junit.Test;
 import static junit.framework.TestCase.*;
 import static org.apache.http.HttpStatus.*;
 
-public class GetOrderByNumberTest {
+public class GetOrderClientByNumberTest {
 
+    public CourierClient courierClient;
     public Courier courier;
-    public VariablesCreate variablesCreate;
     private int courierId;
-    public Order order;
+    public OrderClient orderClient;
 
     @Before
     public void start() {
-        courier = new Courier();
-        variablesCreate = VariablesCreate.getVariablesCreate();
-        courier.createCourier(variablesCreate);
-        courierId = courier.getCourierId(courier, variablesCreate);
-        order = new Order();
+        courierClient = new CourierClient();
+        courier = Courier.getAllVariables();
+        courierClient.createCourier(courier);
+        courierId = courierClient.getCourierId(courierClient, courier);
+        orderClient = new OrderClient();
     }
 
     @Test
     @DisplayName("Checking getting order by number")
     public void canGetOrderByNumberTest() {
-        int orderTrack = order.getOrderTrack(order);
-        Response responseGetOrderByNumber = order.getOrderByNumber(orderTrack);
+        int orderTrack = orderClient.getOrderTrack(orderClient);
+        Response responseGetOrderByNumber = orderClient.getOrderByNumber(orderTrack);
 
         int expectedCode = SC_OK;
         assertEquals("The code should be: " + expectedCode, expectedCode, responseGetOrderByNumber.statusCode());
@@ -38,7 +38,7 @@ public class GetOrderByNumberTest {
     @Test
     @DisplayName("Checking getting order by number without number order")
     public void cannotGetOrderWithoutNumberTest() {
-        Response responseGetOrderByNumber = order.getOrderWithoutNumber();
+        Response responseGetOrderByNumber = orderClient.getOrderWithoutNumber();
 
         int expectedCode = SC_BAD_REQUEST;
         String expectedMessage = "Недостаточно данных для поиска";
@@ -50,7 +50,7 @@ public class GetOrderByNumberTest {
     @Test
     @DisplayName("Checking getting order by number with nonexistent number order")
     public void cannotGetOrderWithNonexistentNumberTest() {
-        Response responseGetOrderByNumber = order.getOrderByNumber(Integer.parseInt(RandomStringUtils.randomNumeric(9)));
+        Response responseGetOrderByNumber = orderClient.getOrderByNumber(Integer.parseInt(RandomStringUtils.randomNumeric(9)));
 
         int expectedCode = SC_NOT_FOUND;
         String expectedMessage = "Заказ не найден";
@@ -61,6 +61,6 @@ public class GetOrderByNumberTest {
 
     @After
     public void finish() {
-        courier.deleteCourier(courierId);
+        courierClient.deleteCourier(courierId);
      }
 }
